fix: add rollback logic + extract shared useAiConsentToggle hook
Resolves the state synchronization bug identified in PR #1171 review.

Problem: handleAiToggle performed consent/withdraw operations then
preference updates. If the second step failed, the contract consent
state would be out of sync with stored preferences.

Solution:
- Extract duplicated handleAiToggle logic into useAiConsentToggle hook
- Use preferences-first-with-rollback pattern:
  1. Update preferences first
  2. Perform consent/withdraw operation
  3. If consent fails, roll back preferences to previous value
- Fixes DRY violation (40+ line function was duplicated across
  PrivacySettingsModal and PrivacySettingsPage)
- Both components now delegate to the shared hook
