+++ apps/learn-card-app/src/pages/privacy-settings/PrivacySettingsPage.tsx
@@ -8,27 +8,21 @@ import {
     useGetPreferencesForDid,
     useUpdatePreferences,
     useGetCurrentLCNUser,
-    useToast,
-    ToastTypeEnum,
 } from 'learn-card-base';
 import { getAiFeatureAgeGateState } from 'learn-card-base';
 import { switchedProfileStore } from 'learn-card-base/stores/walletStore';
 import { useBrandingConfig } from 'learn-card-base/config/TenantConfigProvider';
-import useAutoConsentLearnCardAi from '../../hooks/useAutoConsentLearnCardAi';
-import { useGuardianGate } from '../../hooks/useGuardianGate';
+import { useAiConsentToggle } from '../../hooks/useAiConsentToggle';
 import { useAnalytics } from '../../analytics';
 
 const PrivacySettingsPage: React.FC = () => {
     const history = useHistory();
     const { data: preferences } = useGetPreferencesForDid();
-    const { mutate: updatePreferences, mutateAsync: updatePreferencesAsync } =
+    const { mutate: updatePreferences } =
         useUpdatePreferences();
     const { setEnabled: setAnalyticsEnabled } = useAnalytics();
     const { currentLCNUser } = useGetCurrentLCNUser();
     const brandingConfig = useBrandingConfig();
-    const { presentToast } = useToast();
-    const { guardedAction } = useGuardianGate();
-    const { autoConsentLearnCardAi, withdrawLearnCardAiConsent } = useAutoConsentLearnCardAi();
     const profileType = switchedProfileStore.use.profileType();
 
     // Local DOB fallback so minor banner/locks work even without stored preferences.
@@ -39,6 +33,7 @@ const PrivacySettingsPage: React.FC = () => {
         country: currentLCNUser?.country,
     });
     const isChildProfile = ageGate.isChildProfile;
+    const { handleAiToggle } = useAiConsentToggle({ isChildProfile });
     const isMinor = ageGate.isChildProfile || ageGate.isMinorByAge;
 
     const aiEnabled = ageGate.isAiAgeRestricted
@@ -49,61 +44,7 @@ const PrivacySettingsPage: React.FC = () => {
     const analyticsEnabled = isMinor ? false : preferences?.analyticsEnabled ?? true;
     const bugReportsEnabled = isMinor ? false : preferences?.bugReportsEnabled ?? true;
 
-    const handleAiToggle = useCallback(
-        async (enabled: boolean) => {
-            if (!enabled && isChildProfile) {
-                const withdrawn = await withdrawLearnCardAiConsent();
 
-                if (!withdrawn) {
-                    presentToast('Something went wrong. Please try again.', {
-                        type: ToastTypeEnum.Error,
-                    });
-                    return;
-                }
-
-                await updatePreferencesAsync({ aiEnabled: false });
-                return;
-            }
-
-            if (!enabled || !isChildProfile) {
-                updatePreferences({ aiEnabled: enabled });
-                return;
-            }
-
-            await guardedAction(
-                async () => {
-                    const consented = await autoConsentLearnCardAi({
-                        enabled: true,
-                        userOverrides: {
-                            displayName: currentLCNUser?.displayName ?? '',
-                            image: currentLCNUser?.image ?? '',
-                        },
-                    });
-
-                    if (!consented) {
-                        presentToast('Something went wrong. Please try again.', {
-                            type: ToastTypeEnum.Error,
-                        });
-                        return;
-                    }
-
-                    await updatePreferencesAsync({ aiEnabled: enabled });
-                },
-                { ignorePriorVerification: true }
-            );
-        },
-        [
-            autoConsentLearnCardAi,
-            currentLCNUser?.displayName,
-            currentLCNUser?.image,
-            guardedAction,
-            isChildProfile,
-            presentToast,
-            withdrawLearnCardAiConsent,
-            updatePreferences,
-            updatePreferencesAsync,
-        ]
-    );
 
     const handleAnalyticsToggle = useCallback(
         (enabled: boolean) => {
